Refactor the MailStyle::InlineStyles module
The inlining code has been moved out and is now
completely external to the mailer stuff.
InlineStyles need a new name that is more
appropriate and tests for Inlining are on the way.
